b2bflow_project/requirements.txt
# Dependencies (requirements.txt)
supabase
python-dotenv
httpx